refactor: moves validators to the infra layer
