+++ src/application/validation/index.ts
@@ -1,2 +1,1 @@
 export * from './protocols'
-export * from './validators'
