+++ src/data/use-cases/index.ts
@@ -2,3 +2,4 @@ export * from './db-add-guardian'
 export * from './db-authentication'
 export * from './db-forget-password'
 export * from './db-change-password'
+export * from './forget-password-token-generation'
