+++ src/infra/validators/email-validator-adapter.ts
@@ -1,4 +1,4 @@
-import { type EmailValidator } from '@/application/validation/protocols/email-validator'
+import { type EmailValidator } from '@/application/validation'
 import validator from 'validator'
 
 export class EmailValidatorAdapter implements EmailValidator {
