+++ src/main/factories/controllers/forget-password-factory.ts
@@ -1,9 +1,8 @@
 import { type Controller } from '@/application/protocols'
-import { ForgetPasswordController } from '@/application/controllers/forget-password'
-import { EmailValidatorAdapter } from '@/application/validation/validators'
-import { DbForgetPassword } from '@/data/use-cases'
-import { ForgetPasswordTokenGenerator } from '@/data/use-cases/forget-password-token-generation'
-import { BcryptAdapter } from '@/infra/cryptography/bcrypt-adapter'
+import { ForgetPasswordController } from '@/application/controllers'
+import { EmailValidatorAdapter } from '@/infra/validators'
+import { DbForgetPassword, ForgetPasswordTokenGenerator } from '@/data/use-cases'
+import { BcryptAdapter } from '@/infra/cryptography'
 import { NodeMailerAdapter } from '@/infra/communication'
 import { GuardianAccountRepository, LoggerPgRepository } from '@/infra/repos/postgresql'
 import { LoggerControllerDecorator } from '@/main/decorators'
