+++ tests/src/validation/phone-validator-adapter.spec.ts
@@ -1,4 +1,4 @@
-import { PhoneValidatorAdapter } from '@/application/validation'
+import { PhoneValidatorAdapter } from '@/infra/validators'
 import validator from 'validator'
 
 jest.mock('validator', () => ({
